requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
RPi.GPIO==0.7.1
python-json-logger==2.0.7
asyncio-mqtt==0.16.1
